# docs/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Stemmarest REST API",
    "description" : "Stemmarest is a Neo4j-based repository for variant text traditions (i.e. texts that have been transmitted in multiple manuscripts). For an introduction to its conceptual structure, read on; to dive into specific API calls, see the navigation menu, which is organised by the sort of object each call concerns.\n\n### Definition of terms\n\nWhat follows is a definition of the concepts that occur within Stemmarest; understanding these terms will be helpful for understanding the API documentation (each is also described again, in the same words, at the top of its own section in the navigation menu). Note that most of these entities are defined with respect to (therefore, under) the **/tradition/{tradId}** heading in the API; readings, although defined as root objects, always belong to one and only one tradition section.\n",
    "version" : "1.1"
  },
...
  "paths" : {
...
    "/tradition/{tradId}/relationtype/{name}" : {
      "get" : {
        "tags" : [ "Relation Type" ],
        "summary" : "Get relation type",
        "description" : "Gets the information for the given relation type name.",
        "operationId" : "getRelationType",
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "tradId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "A JSON RelationTypeModel",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RelationTypeModel"
                }
              }
            }
          },
          "400" : {
            "description" : "if the relation type reference is a name shared by multiple types (legacy data only)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "description" : "Not found, if the relation type does not exist",
            "content" : {
              "application/json" : { }
            }
          },
          "500" : {
            "description" : "Failure, with an error report in JSON format",
            "content" : {
              "application/json" : { }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Relation Type" ],
        "summary" : "Create / update relation type specification",
        "description" : "Creates or updates a relation type according to the specification given. If the specification has no name, a new type takes the name in the URL and an existing type keeps its current name. A type that is in use by any relation cannot be renamed.",
        "operationId" : "create_1",
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "tradId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "description" : "A user specification",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RelationTypeModel"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "A JSON RelationTypeModel (existing type was updated)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RelationTypeModel"
                }
              }
            }
          },
          "201" : {
            "description" : "A JSON RelationTypeModel (new type was created)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RelationTypeModel"
                }
              }
            }
          },
          "304" : {
            "description" : "A default type was requested, but could not be generated"
          },
          "400" : {
            "description" : "if the relation type reference is a name shared by multiple types (legacy data only)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "description" : "Not found, if no such tradition exists",
            "content" : {
              "application/json" : { }
            }
          },
          "409" : {
            "description" : "A default type was requested, but a type of that name already exists; the new name conflicts with another existing relation type; or a rename was requested for a type that is still in use by relations",
            "content" : {
              "application/json" : { }
            }
          },
          "500" : {
            "description" : "Failure, with an error report in JSON format",
            "content" : {
              "application/json" : { }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "Relation Type" ],
        "summary" : "Delete a relation type",
        "description" : "Deletes the named relation type.",
        "operationId" : "delete",
        "parameters" : [ {
          "name" : "name",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "tradId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "A JSON RelationTypeModel of the deleted type",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/RelationTypeModel"
                }
              }
            }
          },
          "400" : {
            "description" : "if the relation type reference is a name shared by multiple types (legacy data only)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "404" : {
            "description" : "Not found, if the specified type doesn't exist",
            "content" : {
              "application/json" : { }
            }
          },
          "409" : {
            "description" : "Conflict, if relations of the type still exist in the tradition",
            "content" : {
              "application/json" : { }
            }
          },
          "500" : {
            "description" : "Failure, with an error report in JSON format",
            "content" : {
              "application/json" : { }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "RelationTypeModel" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "defaultsettings" : {
            "type" : "boolean"
          },
          "description" : {
            "type" : "string"
          },
          "display" : {
            "type" : "string"
          },
          "bindlevel" : {
            "type" : "integer",
            "format" : "int32"
          },
          "is_weak" : {
            "type" : "boolean"
          },
          "is_colocation" : {
            "type" : "boolean"
          },
          "is_transitive" : {
            "type" : "boolean"
          },
          "is_generalizable" : {
            "type" : "boolean"
          },
          "use_regular" : {
            "type" : "boolean"
          },
          "name" : {
            "type" : "string"
          }
        }
      },
...
    }
  }
}